New option in Main Menu and Fixed gameplay
Focused on the spawn rate and was able to fix the gun play. It used to be impossible to predict when the shots would actually hit the enemy. I was able to fix that, it was an issue with the gun collider as it was constantly hit by the gun or bullets whenever one would shoot. Spawn rate fixed, used to be spawning it random amount of enemies or at the least more would spawn than needed, so the issue was fixed so that only a certain amount can fix in that given area.