Added Boss phase controller
So I've included the option to have the Boss be able to go into phase 2 or not go into phase 2. It'll give us the option for anyone to place the boss into as a normal enemy but then when he's in his boss area to enable phase 2 to make him harder.